feat: only support actual integers for "INT" attribute type (was double)

+++ src/main/java/tug/tobkul/ontologybrowser/jfxcontroller/attribute/AddAttributePopupController.java
@@ -51,9 +51,9 @@ public class AddAttributePopupController extends AttributePopupController {
                         nameField.getText(),
                         commentField.getText(),
                         AttributeType.INT,
-                        new NumberRangeValue(new NumberRange(Double.parseDouble(minTextField.getText()),
-                                Double.parseDouble(maxTextField.getText()),
-                                Double.parseDouble(intervalTextField.getText())))
+                        new NumberRangeValue(new NumberRange(Integer.parseInt(minTextField.getText()),
+                                Integer.parseInt(maxTextField.getText()),
+                                Integer.parseInt(intervalTextField.getText())))
                 ));
                 attributeTypeChoiceBox.getScene().getWindow().hide();
             }
